+++ src/main/java/com/sparta/bapzip/user/application/dto/request/LoginRequestDto.java
@@ -1,6 +1,5 @@
-package com.sparta.bapzip.user.presentation.dto.request;
+package com.sparta.bapzip.user.application.dto.request;
 
-import com.sparta.bapzip.user.domain.enums.UserRoleEnum;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
@@ -8,18 +7,12 @@ import lombok.Setter;
 
 @Getter
 @Setter
-public class SignupRequestDto {
+public class LoginRequestDto {
 
     @NotBlank
     @Email
     private String email;
 
     @NotBlank
-    private String name;
-
-    @NotBlank
     private String password;
-
-    @NotBlank
-    private UserRoleEnum role;
 }
