+++ src/main/java/com/sparta/bapzip/user/application/dto/request/SignupRequestDto.java
@@ -1,5 +1,6 @@
-package com.sparta.bapzip.user.presentation.dto.request;
+package com.sparta.bapzip.user.application.dto.request;
 
+import com.sparta.bapzip.user.domain.enums.UserRoleEnum;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
@@ -7,12 +8,18 @@ import lombok.Setter;
 
 @Getter
 @Setter
-public class LoginRequestDto {
+public class SignupRequestDto {
 
     @NotBlank
     @Email
     private String email;
 
     @NotBlank
+    private String name;
+
+    @NotBlank
     private String password;
+
+    @NotBlank
+    private UserRoleEnum role;
 }
